refactor: TT-180 move convert_list_to_tuple_string and validate_list to utils
