+++ tests/time_tracker_api/activities/activities_model_test.py
@@ -9,41 +9,6 @@ from time_tracker_api.activities.activities_model import (
 
 
 @pytest.mark.parametrize(
-    "id_list", [123, (1, 2), "id_list", {"id_list": []},],
-)
-def test_validate_list(
-    id_list, activity_repository: ActivityCosmosDBRepository
-):
-    try:
-        activity_repository.validate_list(id_list)
-    except Exception as e:
-        assert type(e) is AssertionError
-
-
-@patch(
-    'time_tracker_api.activities.activities_model.ActivityCosmosDBRepository.validate_list'
-)
-@pytest.mark.parametrize(
-    "id_list,expected_result",
-    [
-        (["id1"], "('id1')"),
-        (["id1", "id2"], "('id1', 'id2')"),
-        (["id1", "id2", "id3", "id4"], "('id1', 'id2', 'id3', 'id4')"),
-    ],
-)
-def test_convert_list_to_tuple_string(
-    validate_list_mock,
-    activity_repository: ActivityCosmosDBRepository,
-    id_list,
-    expected_result,
-):
-    result = activity_repository.convert_list_to_tuple_string(id_list)
-
-    validate_list_mock.assert_called_once()
-    assert expected_result == result
-
-
-@pytest.mark.parametrize(
     "id_list,expected_result",
     [
         (["id1"], "c.id IN ('id1')"),
@@ -52,7 +17,9 @@ def test_convert_list_to_tuple_string(
     ],
 )
 def test_create_sql_in_condition(
-    activity_repository: ActivityCosmosDBRepository, id_list, expected_result,
+    activity_repository: ActivityCosmosDBRepository,
+    id_list,
+    expected_result,
 ):
     result = activity_repository.create_sql_in_condition(id_list)
     assert expected_result == result
