+++ time_tracker_api/activities/activities_model.py
@@ -10,6 +10,7 @@ from commons.data_access_layer.cosmos_db import (
 from time_tracker_api.database import CRUDDao, APICosmosDBDao
 from typing import List, Callable
 from commons.data_access_layer.database import EventContext
+from utils.repository import convert_list_to_tuple_string
 
 
 class ActivityDao(CRUDDao):
@@ -19,7 +20,11 @@ class ActivityDao(CRUDDao):
 container_definition = {
     'id': 'activity',
     'partition_key': PartitionKey(path='/tenant_id'),
-    'unique_key_policy': {'uniqueKeys': [{'paths': ['/name', '/deleted']},]},
+    'unique_key_policy': {
+        'uniqueKeys': [
+            {'paths': ['/name', '/deleted']},
+        ]
+    },
 }
 
 
@@ -51,20 +56,9 @@ class ActivityCosmosDBRepository(CosmosDBRepository):
         )
 
     def create_sql_in_condition(self, id_list):
-        id_values = self.convert_list_to_tuple_string(id_list)
+        id_values = convert_list_to_tuple_string(id_list)
 
         return "c.id IN {value_condition}".format(value_condition=id_values)
-
-    def convert_list_to_tuple_string(self, id_list):
-        self.validate_list(id_list)
-        id_value = (
-            f"('{id_list[0]}')" if len(id_list) == 1 else str(tuple(id_list))
-        )
-        return id_value
-
-    def validate_list(self, id_list):
-        assert isinstance(id_list, list)
-        assert len(id_list) > 0
 
     def find_all_with_id_in_list(
         self,
@@ -85,7 +79,8 @@ class ActivityCosmosDBRepository(CosmosDBRepository):
 
         tenant_id_value = self.find_partition_key_value(event_context)
         result = self.container.query_items(
-            query=query_str, partition_key=tenant_id_value,
+            query=query_str,
+            partition_key=tenant_id_value,
         )
 
         function_mapper = self.get_mapper_or_dict(mapper)
@@ -97,11 +92,13 @@ class ActivityCosmosDBDao(APICosmosDBDao, ActivityDao):
         CosmosDBDao.__init__(self, repository)
 
     def get_all_with_id_in_list(
-        self, id_list,
+        self,
+        id_list,
     ):
         event_ctx = self.create_event_context("read-many")
         activities_list = self.repository.find_all_with_id_in_list(
-            event_ctx, id_list,
+            event_ctx,
+            id_list,
         )
         return activities_list
